[lldb][test] Consolidate libstdc++ and libc++ vector formatter tests into generic test
The libstdc++ test was a subset of the tests in libc++. This test moves
the libc++ test into `generic` and removes the `libstdc++` tests. I
retained the `-D_GLIBCXX_DEBUG` test cases though.

Split out from https://github.com/llvm/llvm-project/pull/146740
