--- lldb/test/API/functionalities/data-formatter/data-formatter-stl/libstdcpp/vector/Makefile
@@ -1,6 +1,0 @@
-CXX_SOURCES := main.cpp
-
-CXXFLAGS := -O0
-USE_LIBSTDCPP := 1
-
-include Makefile.rules
